More Fixes, Getting all Render Sys going
This update is huge and getting out the bad code and fixes to support all suggested rendering systems on suggested OS platforms.

This task is large, harsh and yet as I am getting through it all.

It is not a time atm to pull or download from this system as I need to harshly test, compile and work out the issues and to ensure that all Render Platforms work correctly across each platform and render system.  In short, I am now dealing with the hardest part of this project!

DirectX11 Render system is working and is compliable - verified!

OpenGL Compiles, but nothing is showing (So Yes, errors!)

Vulkan - Compiles shaders, but other than that, not tested or unknown at this moment!

+++ OpenGLCamera.cpp
@@ -966,7 +966,7 @@ void Camera::SetupDefaultCamera(float windowWidth, float windowHeight)
     float fovX_rad = glm::radians(fovX_deg);
     float fovY_rad = 2.0f * atanf(tanf(fovX_rad / 2.0f) / aspect);
 
-    SetProjectionMatrix(glm::perspective(fovY_rad, aspect, config.myConfig.nearPlane, config.myConfig.farPlane));
+    SetProjectionMatrix(glm::perspective(fovY_rad, aspect, static_cast<float>(config.myConfig.nearPlane), static_cast<float>(config.myConfig.farPlane)));
 
     position = eye;
     target   = lookAt;
